PR feedback

+++ src/cast_core/include/cast_core/actors/Loader.hpp
@@ -37,7 +37,7 @@ namespace genny::actor {
 class Loader : public Actor {
 
 public:
-    explicit Loader(ActorContext& context, uint thread, uint totalThreads);
+    explicit Loader(ActorContext& context, uint thread, size_t totalThreads);
     ~Loader() override = default;
 
     static std::string_view defaultName() {
